content_automation/requirements.txt
# Azure Functions
azure-functions==1.18.0

# OpenAI / LLM
openai==1.12.0

# RSS & Web Scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
newspaper3k==0.2.8

# WordPress Integration
python-wordpress-xmlrpc==2.3
requests-toolbelt==1.0.0

# Email Service
sendgrid==6.11.0

# Database
pymysql==1.1.0
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
jinja2==3.1.3
python-dateutil==2.8.2

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
